analyze: refactor: move Rewrite pretty printing tests from expr::hir_op to rewrite module
